+++ tests/test_refresh.py
@@ -28,3 +28,17 @@ def test_saves_timestamp(path_to_created_file, monkeypatch):
     yt_queue.refresh(file)
     data = yt_queue.read(file)
     assert data['refreshed'] > 0
+
+def test_only_if_stale(path_to_created_file, monkeypatch):
+    # when recently refreshed
+    file = path_to_created_file
+    mock_yt_dlp(monkeypatch, extract_info=response_extract_info())
+    yt_queue.refresh(file)
+    data = yt_queue.read(file)
+    original_refreshed_timestamp = data['refreshed']
+
+    # and refreshing again
+    refresh_result = yt_queue.refresh(file, only_if_older='1hour')
+    assert refresh_result is None
+    data = yt_queue.read(file)
+    assert data['refreshed'] == original_refreshed_timestamp